Move game page components to src/components/
