+++ client/src/components/game/cardSizes.scss
@@ -1,0 +1,14 @@
+$card-normal-min-width: 50px;
+$card-normal-min-height: 75px;
+$card-normal-width: 6vw;
+$card-normal-height: 1.5 * $card-normal-width;
+
+$card-big-min-width: 100px;
+$card-big-min-height: 150px;
+$card-big-width: 9vw;
+$card-big-height: 1.5 * $card-big-width;
+
+$reverse-min-width: 75px;
+$reverse-min-height: 112px;
+$reverse-width: 7vw;
+$reverse-height: 1.5 * $reverse-width;
