+++ client/src/components/game/cards/AnswerCard.jsx
@@ -1,3 +1,5 @@
+import './answerCard.scss';
+
 import React from 'react';
 
 const AnswerCard = ({ text }) => (
