+++ client/src/components/game/cards/QuestionCard.jsx
@@ -1,3 +1,5 @@
+import './questionCard.scss';
+
 import React from 'react';
 
 const QuestionCard = ({ text }) => (
