+++ client/src/components/game/cards/answerCard.scss
@@ -1,0 +1,16 @@
+@import '../../../shared/styles/globals.scss';
+@import '../cardSizes.scss';
+
+.answer-card {
+  width: $card-normal-width;
+  height: $card-normal-height;
+  min-width: $card-normal-min-width;
+  min-height: $card-normal-min-height;
+
+  background-color: $white;
+  margin: 0 10px;
+
+  span {
+    color: $black;
+  }
+}
