+++ client/src/components/game/containers/cardsInHand.scss
@@ -1,0 +1,8 @@
+.cards-in-hand {
+  grid-column: 2 / 2;
+  grid-row: 3 / 3;
+
+  width: 100%;
+  display: flex;
+  justify-content: center;
+}
