+++ client/src/components/game/containers/index.js
@@ -1,0 +1,2 @@
+export { CardsInHand } from './CardsInHand';
+export { PlayedCards } from './PlayedCards';
