+++ client/src/components/game/counter/counter.scss
@@ -1,0 +1,6 @@
+@import '../../../shared/styles/globals.scss';
+
+.counter {
+  font-size: 2rem;
+  color: $black;
+}
